#75[MAINTENANCE]: 코드 분리 작업
