+++ Poki-iOS/Poki-iOS/Controllers/SettingVC/NoticeListViewController.swift
@@ -14,33 +14,14 @@ class NoticeListViewController: UIViewController {
     // MARK: - Properties
     
     private let tableView = UITableView()
-
-    private let noticeItems: [NoticeItem] = [
-        NoticeItem(title: "중요한 공지사항", date: "2023.10.20"),
-        NoticeItem(title: "새로운 업데이트 안내", date: "2023.10.18"),
-        NoticeItem(title: "이벤트 참여 방법", date: "2023.10.15"),
-        NoticeItem(title: "긴 공지사항 입니다. 긴 공지사항입니다. 긴 공지사항 입니다. 긴 공지사항입니다. 긴 공지사항 입니다. 긴 공지사항입니다. 긴 공지사항 입니다. 긴 공지사항입니다.", date: "2023.10.20")
-    ]
-    
+    private var noticeItems: [Notice] = NoticeData.noticeItems
     
     // MARK: - Life Cycle
     
     override func viewDidLoad() {
         super.viewDidLoad()
         view.backgroundColor = .white
-        
-        tableView.dataSource = self
-        tableView.delegate = self
-        tableView.register(NoticeListTableViewCell.self, forCellReuseIdentifier: "NoticeListTableViewCell")
-        
-        view.addSubview(tableView)
-        
-        tableView.snp.makeConstraints { make in
-            make.top.equalTo(view.safeAreaLayoutGuide).offset(20)
-            make.leading.equalToSuperview().offset(10)
-            make.trailing.equalToSuperview().offset(-10)
-            make.bottom.equalToSuperview().offset(-10)
-        }
+        setupTableView()
     }
     
     override func viewWillAppear(_ animated: Bool) {
@@ -73,18 +54,22 @@ class NoticeListViewController: UIViewController {
         navigationController?.navigationBar.scrollEdgeAppearance = appearance
     }
     
-    
-    
-    
-    
-    
+    private func setupTableView() {
+        tableView.dataSource = self
+        tableView.delegate = self
+        tableView.register(NoticeListTableViewCell.self, forCellReuseIdentifier: "NoticeListTableViewCell")
+        
+        view.addSubview(tableView)
+        
+        tableView.snp.makeConstraints { make in
+            make.top.equalTo(view.safeAreaLayoutGuide).offset(20)
+            make.leading.equalToSuperview().offset(10)
+            make.trailing.equalToSuperview().offset(-10)
+            make.bottom.equalToSuperview().offset(-10)
+        }
+    }
     // MARK: - Actions
-    
-    
-    
-    
-    
-    
+
 }
 
 // MARK: - UITableViewDataSource, UITableViewDelegate
@@ -107,54 +92,5 @@ extension NoticeListViewController: UITableViewDataSource, UITableViewDelegate {
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
         return 90
     }
-    
-    
 }
-
-// MARK: - CustomCell
-
-class NoticeListTableViewCell: UITableViewCell {
-    let titleLabel: UILabel = {
-        let label = UILabel()
-        label.font = UIFont(name: Constants.font, size: 14)
-        label.numberOfLines = 2 // 최대 2줄 표시
-        return label
-    }()
-    
-    let dateLabel: UILabel = {
-        let label = UILabel()
-        label.font = UIFont(name: Constants.font, size: 10)
-        label.textColor = .gray
-        return label
-    }()
-    
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        
-        contentView.addSubview(titleLabel)
-        contentView.addSubview(dateLabel)
-        
-        titleLabel.snp.makeConstraints { make in
-            make.top.left.equalTo(contentView).offset(16)
-            make.right.equalTo(contentView).offset(-16)
-        }
-        
-        dateLabel.snp.makeConstraints { make in
-            make.top.equalTo(titleLabel.snp.bottom).offset(8)
-            make.left.equalTo(contentView).offset(16)
-            make.right.equalTo(contentView).offset(-16)
-            make.bottom.equalTo(contentView).offset(-16)
-        }
-    }
-    
-    required init?(coder aDecoder: NSCoder) {
-        super.init(coder: aDecoder)
-    }
-}
-
-struct NoticeItem {
-    var title: String
-    var date: String
-}
-
